refactor(ui): Group 2FA setup components into a dedicated folder
